feat(messaging): remove campaign Overview tab in favor of workflow editor
