+++ products/messaging/frontend/Campaigns/campaignSceneLogic.ts
@@ -8,7 +8,7 @@ import { Breadcrumb } from '~/types'
 
 import type { campaignSceneLogicType } from './campaignSceneLogicType'
 
-export const CampaignTabs = ['overview', 'workflow', 'logs', 'metrics'] as const
+export const CampaignTabs = ['workflow', 'logs', 'metrics'] as const
 export type CampaignTab = (typeof CampaignTabs)[number]
 
 export interface CampaignSceneLogicProps {
@@ -24,7 +24,7 @@ export const campaignSceneLogic = kea<campaignSceneLogicType>([
     }),
     reducers({
         currentTab: [
-            'overview' as CampaignTab,
+            'workflow' as CampaignTab,
             {
                 setCurrentTab: (_, { tab }) => tab,
             },
