+++ products/messaging/frontend/Campaigns/hogflows/steps/StepTrigger.tsx
@@ -1,12 +1,16 @@
 import { Node } from '@xyflow/react'
 import { useActions, useValues } from 'kea'
+import posthog from 'posthog-js'
 
-import { IconBolt, IconWebhooks } from '@posthog/icons'
-import { LemonLabel, LemonSelect } from '@posthog/lemon-ui'
+import { IconBolt, IconPlusSmall, IconWebhooks } from '@posthog/icons'
+import { LemonButton, LemonDivider, LemonLabel, LemonSelect, LemonTag, lemonToast } from '@posthog/lemon-ui'
 
 import { CodeSnippet } from 'lib/components/CodeSnippet'
+import { PropertyFilters } from 'lib/components/PropertyFilters/PropertyFilters'
+import { TaxonomicFilterGroupType } from 'lib/components/TaxonomicFilter/types'
 import { useFeatureFlag } from 'lib/hooks/useFeatureFlag'
 import { LemonField } from 'lib/lemon-ui/LemonField'
+import { LemonRadio } from 'lib/lemon-ui/LemonRadio'
 import { publicWebhooksHostOrigin } from 'lib/utils/apiHost'
 
 import { campaignLogic } from '../../campaignLogic'
@@ -112,6 +116,11 @@ function StepTriggerConfigurationEvents({
                     buttonCopy="Add trigger event"
                 />
             </LemonField.Pure>
+
+            <LemonDivider />
+            <ConversionGoalSection />
+            <LemonDivider />
+            <ExitConditionSection />
         </>
     )
 }
@@ -161,5 +170,115 @@ function StepTriggerConfigurationWebhook({
                 errors={validationResult?.errors}
             />
         </>
+    )
+}
+
+function ConversionGoalSection(): JSX.Element {
+    const { setCampaignValue } = useActions(campaignLogic)
+    const { campaign } = useValues(campaignLogic)
+
+    return (
+        <div className="flex flex-col py-2 w-full">
+            <span className="text-md font-semibold">Conversion goal (optional)</span>
+            <p>Define what a user must do to be considered converted.</p>
+
+            <div className="flex gap-1 max-w-240">
+                <div className="flex flex-col flex-2 gap-4">
+                    <LemonField.Pure label="Detect conversion from property changes">
+                        <PropertyFilters
+                            buttonText="Add property conversion"
+                            propertyFilters={campaign.conversion?.filters ?? []}
+                            taxonomicGroupTypes={[
+                                TaxonomicFilterGroupType.PersonProperties,
+                                TaxonomicFilterGroupType.Cohorts,
+                                TaxonomicFilterGroupType.HogQLExpression,
+                            ]}
+                            onChange={(filters) => setCampaignValue('conversion', { ...campaign.conversion, filters })}
+                            pageKey="campaign-conversion-properties"
+                            hideBehavioralCohorts
+                        />
+                    </LemonField.Pure>
+                    <div className="flex flex-col gap-1">
+                        <LemonLabel>
+                            Detect conversion from events
+                            <LemonTag>Coming soon</LemonTag>
+                        </LemonLabel>
+                        <LemonButton
+                            type="secondary"
+                            size="small"
+                            icon={<IconPlusSmall />}
+                            onClick={() => {
+                                posthog.capture('messaging campaign event conversion clicked')
+                                lemonToast.info('Event targeting coming soon!')
+                            }}
+                        >
+                            Add event conversion
+                        </LemonButton>
+                    </div>
+                </div>
+                <LemonDivider vertical />
+                <div className="flex-1">
+                    <LemonField.Pure
+                        label="Conversion window"
+                        info="How long after entering the campaign should we check for conversion? After this window, users will be considered for conversion."
+                    >
+                        <LemonSelect
+                            value={campaign.conversion?.window_minutes}
+                            onChange={(value) =>
+                                setCampaignValue('conversion', {
+                                    ...campaign.conversion,
+                                    window_minutes: value,
+                                })
+                            }
+                            placeholder="No conversion window"
+                            allowClear
+                            options={[
+                                { value: 24 * 60 * 60, label: '24 hours' },
+                                { value: 7 * 24 * 60 * 60, label: '7 days' },
+                                { value: 14 * 24 * 60 * 60, label: '14 days' },
+                                { value: 30 * 24 * 60 * 60, label: '30 days' },
+                            ]}
+                        />
+                    </LemonField.Pure>
+                </div>
+            </div>
+        </div>
+    )
+}
+
+function ExitConditionSection(): JSX.Element {
+    const { setCampaignValue } = useActions(campaignLogic)
+    const { campaign } = useValues(campaignLogic)
+
+    return (
+        <div className="flex flex-col flex-1 w-full py-2">
+            <span className="text-md font-semibold">Exit condition</span>
+            <p>Choose how your users move through the campaign.</p>
+
+            <LemonField.Pure>
+                <LemonRadio
+                    value={campaign.exit_condition ?? 'exit_only_at_end'}
+                    onChange={(value) => setCampaignValue('exit_condition', value)}
+                    options={[
+                        {
+                            value: 'exit_only_at_end',
+                            label: 'Exit at end of workflow',
+                        },
+                        {
+                            value: 'exit_on_trigger_not_matched',
+                            label: 'Exit on trigger not matched',
+                        },
+                        {
+                            value: 'exit_on_conversion',
+                            label: 'Exit on conversion',
+                        },
+                        {
+                            value: 'exit_on_trigger_not_matched_or_conversion',
+                            label: 'Exit on trigger not matched or conversion',
+                        },
+                    ]}
+                />
+            </LemonField.Pure>
+        </div>
     )
 }
